requirements.txt
tika
requests
pandas~=1.1.3
numpy~=1.19.2
pytz~=2020.1
newspaper3k
